+++ manuals/refman/refman.sh
@@ -125,6 +125,7 @@ eval $xslt_proc -o predicates/logtalk_library_path_2.section refman.xsl predicat
 eval $xslt_proc -o predicates/logtalk_load_context_2.section refman.xsl predicates/logtalk_load_context_2.html
 eval $xslt_proc -o predicates/current_logtalk_flag_2.section refman.xsl predicates/current_logtalk_flag_2.html
 eval $xslt_proc -o predicates/set_logtalk_flag_2.section refman.xsl predicates/set_logtalk_flag_2.html
+eval $xslt_proc -o predicates/create_logtalk_flag_3.section refman.xsl predicates/create_logtalk_flag_3.html
 eval $xslt_proc -o methods/parameter_2.section refman.xsl methods/parameter_2.html
 eval $xslt_proc -o methods/self_1.section refman.xsl methods/self_1.html
 eval $xslt_proc -o methods/sender_1.section refman.xsl methods/sender_1.html
@@ -260,6 +261,7 @@ cat -s \
 	predicates/logtalk_load_context_2.section \
 	predicates/current_logtalk_flag_2.section \
 	predicates/set_logtalk_flag_2.section \
+	predicates/create_logtalk_flag_3.section \
 	methods.header \
 	methods/parameter_2.section \
 	methods/self_1.section \